requirements.txt
# core
crewai>=0.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.3
langchain-openai>=0.0.8
langchain-groq>=1.1.0
langchain-text-splitters>=0.0.1

# vector store
faiss-cpu>=1.7.4

# embeddings
sentence-transformers>=2.3.1

# document loading
unstructured>=0.12.0

# env + config
python-dotenv>=1.0.0
pyyaml>=6.0.1

# data validation
pydantic>=2.0.0

# utilities
tqdm>=4.66.0
rich>=13.7.0

# eval
pandas>=2.0.0
tabulate>=0.9.0

# api
flask>=3.0.0
flask-cors>=4.0.0

# optional
openai>=1.12.0